Rename fragments file to baseQuery
